pc-server/requirements.txt
pynput>=1.7.6
pyusb>=1.2.1
